Virtual Kubelet: configure custom rate limiter
By default, the virtual kubelet uses the DefaultControllerRateLimiter
configuration for the workqueues, composed of an overall bucket rate
limiter and a per-item exponential rate limiter to address failures.
Though, the former introduces performance limitations when handling a
high number of pods. This commit configures a custom RateLimiter,
composed only of the failure backoff part, to prevent that burden.