Fixed network manager panic due to configuration inconsistency
This PR fixes an issue related to the network manager. It is possible that the network manager is killed or re-scheduled during the termination of an endpoint mapping. In this case, there could be inconsistencies in the IPAM configuration, in particular, further requests of terminating the same mapping, could result in panicing because the IP has been already freed, but this do not appear in some structures of IPAM module. The release of an IP has been made to be idempotent, so that further requests won't panic.