fix (network-manager): correct a bug (ref #912) which prevented the network manager to restart
Subnets in CIDR notation can be passed to the network manager as parameters
in order to prevent them to be used for remote clusters. A bug prevented
the network manager to restart when networks to be reserved are passed as
arguments.  The bug introduced in PR #868 has been fixed. If a subnet has to be
removed from the reserved networks we just remove the parameter from the network
manager and restart the operator.